S3 filters

+++ api/Domain/Migrations/20200803025653_EcommInit.cs
@@ -40,7 +40,7 @@ namespace Domain.Migrations
                         .Annotation("Sqlite:Autoincrement", true),
                     Name = table.Column<string>(nullable: true),
                     Description = table.Column<string>(nullable: true),
-                    Price = table.Column<decimal>(nullable: false),
+                    Price = table.Column<double>(nullable: false),
                     PictureUrl = table.Column<string>(nullable: true),
                     ProductCategoryId = table.Column<int>(nullable: false),
                     ProductBrandId = table.Column<int>(nullable: false)
